fix: test file location and fix test_html_hyperlink_parse
