+++ test/Cargo.toml
@@ -7,4 +7,5 @@ publish = false
 
 [dependencies]
 shiva = { path = "../lib" }
+serde_json = "1.0"
 anyhow = "1.0.75"
